fix: use read stream to read the file
