+++ csv/file.csv
@@ -1,3 +1,6 @@
-fieldA.title, fieldA.children.0.name, fieldA.children.0.id,fieldA.children.1.name, fieldA.children.1.employee.0.name,fieldA.children.1.employee.1.name, fieldA.address.0,fieldA.address.1, description
-Food Factory, Oscar, 0023, Tikka, Tim, Joe, 3 Lame Road, Grantstown, A fresh new food factory
-Kindom Garden, Ceil, 54, Pillow, Amst, Tom, 24 Shaker Street, HelloTown, Awesome castle
+Book,Author,Amount,Price
+The Compound Effect,Darren Hardy,5,9.48
+The 7 Habits of Highly Effective People,Stephen R. Covey,4,23.48
+The Miracle Morning,Hal Elrod,10,21.34
+Influence: The Psychology of Persuasion,Robert B. Cialdini,4,12.99
+The ONE Thing,Gary Keller,1,11.18
