+++ index.js
@@ -1,1 +1,2 @@
+require('./task1');
 require('./task3');
